Split seperation of BLE started
